Stop using Start-Process in setup-distro action
The failure of the resulting process is not resulting in the failure of
the powershell command. There are more instances of this in other
actions.

Refactor how /etc/wsl.conf, /etc/sudoers.d/user.conf and
/srv/github-worskpace are created to simplify quoting and append
interacting with powershell.

Writing to Linux files from Windows needs to be done carefully, to
avoid using CR-LF line endings that confuse some of the software.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>